+++ NEWS.md
@@ -1,3 +1,9 @@
+# chromatographR 0.3.1
+
+* Added support for variable penalty dynamic time warping in `correct_RT`.
+* Allow use of raw data for peak integration in `get_peaks`.
+* Added `verbose` option to `correct_rt` to print reference chromatogram.
+
 # chromatographR 0.3
 
 * Added a `NEWS.md` file to track changes to the package.